Move connections grid to main dashboard view
